refactor: reorganize code into modules
- Move chat-related code to src/chat/index.ts
- Move database-related code to src/database/index.ts
- Move cache-related code to src/cache/index.ts
- Move clients-related code to src/clients/index.ts
- Move config-related code to src/config/index.ts
- Update src/index.ts to use these modules

This refactoring improves code organization by separating concerns into dedicated modules while maintaining existing functionality.
