+++ src/clients/index.ts
@@ -1,0 +1,44 @@
+import { AutoClientInterface } from "@ai16z/client-auto";
+import { DiscordClientInterface } from "@ai16z/client-discord";
+import { TelegramClientInterface } from "@ai16z/client-telegram";
+import { TwitterClientInterface } from "@ai16z/client-twitter";
+import type { Character, IAgentRuntime } from "@ai16z/eliza";
+
+export async function initializeClients(
+  character: Character,
+  runtime: IAgentRuntime
+) {
+  const clients = [];
+  const clientTypes = character.clients?.map((str) => str.toLowerCase()) || [];
+
+  if (clientTypes.includes("auto")) {
+    const autoClient = await AutoClientInterface.start(runtime);
+    if (autoClient) clients.push(autoClient);
+  }
+
+  if (clientTypes.includes("discord")) {
+    clients.push(await DiscordClientInterface.start(runtime));
+  }
+
+  if (clientTypes.includes("telegram")) {
+    const telegramClient = await TelegramClientInterface.start(runtime);
+    if (telegramClient) clients.push(telegramClient);
+  }
+
+  if (clientTypes.includes("twitter")) {
+    const twitterClients = await TwitterClientInterface.start(runtime);
+    clients.push(twitterClients);
+  }
+
+  if (character.plugins?.length > 0) {
+    for (const plugin of character.plugins) {
+      if (plugin.clients) {
+        for (const client of plugin.clients) {
+          clients.push(await client.start(runtime));
+        }
+      }
+    }
+  }
+
+  return clients;
+}
